Moved dispatcher provider to rpc package
